runtime: move inject queue to `tokio::runtime::task` (#3939)
